% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hana.R
\docType{data}
\name{hana}
\alias{hana}
\title{Hana written by Akutagawa Ryunosuke}
\format{
An object of class \code{character} of length 161.
}
\source{
\url{https://www.aozora.gr.jp/cards/000879/files/42_ruby_154.zip}
}
\usage{
hana
}
\description{
Japanese text sample. The text originally came from Aozora Bunko.
}
\keyword{datasets}
