% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install.packages}
\alias{install.packages}
\title{Mask over utils::install.packages}
\usage{
install.packages(..., .opts = "--no-multiarch")
}
\arguments{
\item{...}{Passed to \code{utils::install.packages}.}

\item{.opts}{Install options.}
}
\description{
Mask over utils::install.packages
}
