% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nano}
\alias{nano}
\title{An R implementation of NanoID}
\usage{
nano(size = 21L, alpha = c(LETTERS, letters, as.character(0:9), "_", "-"))
}
\arguments{
\item{size}{Integer.}
}
\value{
string scalar.
}
\description{
An R implementation of NanoID
}
