% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silicon.R
\name{silicon}
\alias{silicon}
\title{Execute Silicon Command}
\usage{
silicon(
  input_file = NULL,
  flags = list(no_line_number = FALSE, no_round_corner = FALSE, no_window_controls =
    FALSE, to_clipboard = TRUE),
  language = "R",
  output_path = getwd(),
  ...
)
}
\arguments{
\item{input_file}{path to the source code. If provided `NULL`, render from clipboard.}

\item{flags}{list of flag sets.}

\item{language}{default value is "R".}

\item{output_path}{where the output is created. If `flags$to_clipboard` is TRUE, this is ignored.}

\item{...}{other arguments are passed to silicon's options as `args` of \code{processx::run("silicon")}.}
}
\value{
result from \code{processx::run("silicon")} invisibly.
}
\description{
Execute Silicon Command
}
