% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siliconate-package.R
\docType{package}
\name{siliconate}
\alias{siliconate}
\alias{siliconate-package}
\title{siliconate: Miscellaneous Functions of Paithiov909}
\description{
My own miscellaneous functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/siliconate}
  \item Report bugs at \url{https://github.com/paithiov909/siliconate/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
