% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_locale}
\alias{switch_locale}
\title{Switch locale betwenn English and Japanese}
\usage{
switch_locale(lang = c("en", "ja"))
}
\arguments{
\item{lang}{Choose one of 'en' or 'ja'}
}
\value{
The settled locale setting is returned invisibly.
}
\description{
Switch locale betwenn English and Japanese
}
