% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_pkgtype}
\alias{switch_pkgtype}
\title{Switch pkgType to 'binary'}
\usage{
switch_pkgtype(pkg_type = c("binary", "source"))
}
\arguments{
\item{pkg_type}{Choose one of 'binary' or 'source'}
}
\value{
The settled pktType option is returned invisibly.
}
\description{
Switch pkgType option to `binary` if possible.
}
